requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
sqlalchemy==1.4.46
psycopg2-binary==2.9.5
python-multipart==0.0.6
python-dotenv==0.21.1
PyPDF2==3.0.1
pydantic==1.10.2
alembic==1.9.4
requests==2.28.2
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==8.3.2 